Add page redirects for logged in users
If a user is logged in as a guest or admin and attempts to visit one of
the login pages, they get redirected either to the homepage or the admin
dashboard.

A user who isn't in an admin session will be redirected to the homepage
when they try to visit the admin dashboard.